requirements.txt
bidict==0.21.2
certifi==2021.5.30
chardet==4.0.0
idna==2.10
lxml==4.6.3
multitasking==0.0.9
numpy==1.21.0
pandas==1.3.0
plotly==5.1.0
python-dateutil==2.8.1
python-engineio==4.2.0
python-socketio==5.3.0
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlite-spellfix==1.0
tenacity==7.0.0
urllib3==1.26.6
yfinance==0.1.60
